requirements.txt
streamlit
pandas
numpy
plotly
matplotlib
biopython
requests
streamlit-agraph
networkx
google-generativeai
python-dotenv
